test(User): 사용자 입력 테스트 -> 3자리 숫자입력, 재시작 숫자입력 2개로 분리
